Initial "fuzz" command.

+++ src/main/java/com/powergem/wce/entities/LineCostDatumEntity.java
@@ -1,4 +1,4 @@
 package com.powergem.wce.entities;
 
-public record LineCostDatumEntity(int scenarioId, int id, float length, float maxRatingPerLine, float maxAllowedFlowPerLine, float upgradeCost, float newLineCost) {
+public record LineCostDatumEntity(int scenarioId, int id, double length, double maxRatingPerLine, double maxAllowedFlowPerLine, double upgradeCost, double newLineCost) {
 }
